+++ qiskit/ignis/verification/tomography/experiment/__init__.py
@@ -1,0 +1,2 @@
+from .process_tomography import ProcessTomographyExperiment
+from .state_tomography import StateTomographyExperiment
